test: update comments
