+++ packages/taiko-client/internal/testutils/helper.go
@@ -5,19 +5,29 @@ import (
 	"crypto/ecdsa"
 	"crypto/rand"
 	"math/big"
+	"os"
 	"time"
 
 	"github.com/cenkalti/backoff/v4"
+	"github.com/ethereum-optimism/optimism/op-service/txmgr"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/beacon/engine"
 	"github.com/ethereum/go-ethereum/common"
+	consensus "github.com/ethereum/go-ethereum/consensus/taiko"
+	"github.com/ethereum/go-ethereum/core/rawdb"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/params"
+	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/phayes/freeport"
 
+	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/encoding"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/metadata"
 	pacayaBindings "github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/pacaya"
+	anchortxconstructor "github.com/taikoxyz/taiko-mono/packages/taiko-client/driver/anchor_tx_constructor"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/rpc"
+	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/utils"
 )
 
 func (s *ClientTestSuite) proposeEmptyBlockOp(ctx context.Context, proposer Proposer) {
@@ -311,4 +321,106 @@ func SendDynamicFeeTx(
 		return nil, err
 	}
 	return tx, nil
+}
+
+func (s *ClientTestSuite) resetToBaseBlock(key *ecdsa.PrivateKey) {
+	// Propose an empty block at genesis.
+	t := s.TxMgr("proposer", key)
+	inbox := common.HexToAddress(os.Getenv("PROVER_SET"))
+	params := &encoding.BatchParams{
+		Proposer:   crypto.PubkeyToAddress(key.PublicKey),
+		Coinbase:   common.HexToAddress(RandomHash().Hex()),
+		BlobParams: encoding.BlobParams{ByteOffset: 0, ByteSize: 0},
+		Blocks: []pacayaBindings.ITaikoInboxBlockParams{{
+			NumTransactions: 0,
+			TimeShift:       0,
+			SignalSlots:     make([][32]byte, 0),
+		}},
+	}
+	encoded, err := encoding.EncodeBatchParamsWithForcedInclusion(nil, params)
+	s.Nil(err)
+
+	emptyTxlistBytes, err := utils.EncodeAndCompressTxList(types.Transactions{})
+	s.Nil(err)
+
+	data, err := encoding.TaikoInboxABI.Pack("proposeBatch", encoded, emptyTxlistBytes)
+	s.Nil(err)
+
+	sink := make(chan *pacayaBindings.TaikoInboxClientBatchProposed, 1)
+	sub, err := s.RPCClient.PacayaClients.TaikoInbox.WatchBatchProposed(nil, sink)
+	s.Nil(err)
+	defer func() {
+		sub.Unsubscribe()
+		close(sink)
+	}()
+	_, err = t.Send(
+		context.Background(),
+		txmgr.TxCandidate{TxData: data, To: &inbox, Blobs: nil, GasLimit: 1_000_000},
+	)
+	s.Nil(encoding.TryParsingCustomError(err))
+	e := <-sink
+	s.NotNil(e)
+
+	// After received the event, we start building the payload.
+	difficulty, err := encoding.CalculatePacayaDifficulty(new(big.Int).SetUint64(e.Info.LastBlockId))
+	s.Nil(err)
+
+	parent, err := s.RPCClient.L2.HeaderByNumber(
+		context.Background(),
+		new(big.Int).Sub(new(big.Int).SetUint64(e.Info.LastBlockId), common.Big1),
+	)
+	s.Nil(err)
+
+	baseFee, err := s.RPCClient.CalculateBaseFee(
+		context.Background(), parent, &e.Info.BaseFeeConfig, e.Info.LastBlockTimestamp,
+	)
+	s.Nil(err)
+
+	anchorBlock, err := s.RPCClient.L1.HeaderByNumber(context.Background(), new(big.Int).SetUint64(e.Info.AnchorBlockId))
+	s.Nil(err)
+
+	anchorTxConstructor, err := anchortxconstructor.New(s.RPCClient)
+	s.Nil(err)
+	anchor, err := anchorTxConstructor.AssembleAnchorV3Tx(
+		context.Background(),
+		anchorBlock.Number,
+		anchorBlock.Root,
+		parent,
+		&e.Info.BaseFeeConfig,
+		[][32]byte{},
+		new(big.Int).SetUint64(e.Info.LastBlockId),
+		baseFee,
+	)
+	s.Nil(err)
+
+	txListBytes, err := rlp.EncodeToBytes(types.Transactions{anchor})
+	s.Nil(err)
+
+	// Call engine APIs to insert the base block.
+	s.forkTo(&engine.PayloadAttributes{
+		Timestamp:             e.Info.LastBlockTimestamp,
+		Random:                common.BytesToHash(difficulty),
+		SuggestedFeeRecipient: e.Info.Coinbase,
+		Withdrawals:           []*types.Withdrawal{},
+		BlockMetadata: &engine.BlockMetadata{
+			Beneficiary: e.Info.Coinbase,
+			GasLimit:    uint64(e.Info.GasLimit) + consensus.AnchorV3GasLimit,
+			Timestamp:   e.Info.LastBlockTimestamp,
+			TxList:      txListBytes,
+			MixHash:     common.Hash(difficulty),
+			ExtraData:   e.Info.ExtraData[:],
+		},
+		BaseFeePerGas: baseFee,
+		L1Origin: &rawdb.L1Origin{
+			BlockID:            new(big.Int).SetUint64(e.Info.LastBlockId),
+			L1BlockHeight:      new(big.Int).SetUint64(e.Raw.BlockNumber),
+			L2BlockHash:        common.Hash{},
+			L1BlockHash:        e.Raw.BlockHash,
+			BuildPayloadArgsID: [8]byte{},
+		},
+	}, parent.Hash())
+
+	head, err := s.RPCClient.L2.HeaderByNumber(context.Background(), nil)
+	s.Nil(err)
+	s.Equal(common.Big1, head.Number)
 }
