+++ packages/taiko-client/internal/testutils/suite.go
@@ -14,7 +14,6 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/beacon/engine"
 	"github.com/ethereum/go-ethereum/common"
-	consensus "github.com/ethereum/go-ethereum/consensus/taiko"
 	"github.com/ethereum/go-ethereum/core/rawdb"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
@@ -23,8 +22,6 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/encoding"
-	pacayaBindings "github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/pacaya"
-	anchortxconstructor "github.com/taikoxyz/taiko-mono/packages/taiko-client/driver/anchor_tx_constructor"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/jwt"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/rpc"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/utils"
@@ -97,9 +94,10 @@ func (s *ClientTestSuite) SetupTest() {
 		s.depositProverSetTokens(ownerPrivKey)
 	}
 
+	// At the beginning of each test, we ensure the L2 chain has been reset to the base block (height: 1).
 	s.once.Do(func() {
 		s.testnetL1SnapshotID = s.SetL1Snapshot()
-		s.initBaseBlock(l1ProposerPrivKey)
+		s.resetToBaseBlock(l1ProposerPrivKey)
 	})
 
 	s.BlobServer = NewMemoryBlobServer()
@@ -224,7 +222,7 @@ func (s *ClientTestSuite) KeyFromEnv(envName string) *ecdsa.PrivateKey {
 func (s *ClientTestSuite) TearDownTest() {
 	s.RevertL1Snapshot(s.testnetL1SnapshotID)
 	s.testnetL1SnapshotID = s.SetL1Snapshot()
-	s.initBaseBlock(s.KeyFromEnv("L1_PROPOSER_PRIVATE_KEY"))
+	s.resetToBaseBlock(s.KeyFromEnv("L1_PROPOSER_PRIVATE_KEY"))
 	_, err := s.RPCClient.L2Engine.SetHeadL1Origin(context.Background(), common.Big1)
 	s.Nil(err)
 	s.BlobServer.Close()
@@ -241,6 +239,10 @@ func (s *ClientTestSuite) SetHead(headNum *big.Int) {
 		return
 	}
 
+	// For other nodes, we need to use the engine API to set the head instead,
+	// to reset the chain head to a block which is already in the canonical chain,
+	// we need to fork the chain to a block with different attributes at the same height
+	// at first, then set the canonical head to the block we want.
 	block, err := s.RPCClient.L2.BlockByNumber(context.Background(), headNum)
 	s.Nil(err)
 
@@ -273,11 +275,12 @@ func (s *ClientTestSuite) SetHead(headNum *big.Int) {
 			BuildPayloadArgsID: [8]byte{},
 		},
 	}
+	// Set the chain head to a block with different attributes at first.
 	attributes.SuggestedFeeRecipient = common.HexToAddress(RandomHash().Hex())
 	attributes.BlockMetadata.Beneficiary = attributes.SuggestedFeeRecipient
-
 	s.forkTo(attributes, block.ParentHash())
 
+	// Set the chain head back to the block we want.
 	attributes.SuggestedFeeRecipient = originalCoinbase
 	attributes.BlockMetadata.Beneficiary = originalCoinbase
 	s.forkTo(attributes, block.ParentHash())
@@ -351,103 +354,4 @@ func (s *ClientTestSuite) forkTo(attributes *engine.PayloadAttributes, parentHas
 	s.Nil(err)
 
 	s.Equal(attributes.L1Origin.BlockID.Uint64(), head.Number.Uint64())
-}
-
-func (s *ClientTestSuite) initBaseBlock(key *ecdsa.PrivateKey) {
-	t := s.TxMgr("proposer", key)
-	inbox := common.HexToAddress(os.Getenv("PROVER_SET"))
-	params := &encoding.BatchParams{
-		Proposer:   crypto.PubkeyToAddress(key.PublicKey),
-		Coinbase:   common.HexToAddress(RandomHash().Hex()),
-		BlobParams: encoding.BlobParams{ByteOffset: 0, ByteSize: 0},
-		Blocks: []pacayaBindings.ITaikoInboxBlockParams{{
-			NumTransactions: 0,
-			TimeShift:       0,
-			SignalSlots:     make([][32]byte, 0),
-		}},
-	}
-	encoded, err := encoding.EncodeBatchParamsWithForcedInclusion(nil, params)
-	s.Nil(err)
-
-	emptyTxlistBytes, err := utils.EncodeAndCompressTxList(types.Transactions{})
-	s.Nil(err)
-
-	data, err := encoding.TaikoInboxABI.Pack("proposeBatch", encoded, emptyTxlistBytes)
-	s.Nil(err)
-
-	sink := make(chan *pacayaBindings.TaikoInboxClientBatchProposed, 1)
-	sub, err := s.RPCClient.PacayaClients.TaikoInbox.WatchBatchProposed(nil, sink)
-	s.Nil(err)
-	defer func() {
-		sub.Unsubscribe()
-		close(sink)
-	}()
-	_, err = t.Send(
-		context.Background(),
-		txmgr.TxCandidate{TxData: data, To: &inbox, Blobs: nil, GasLimit: 1_000_000},
-	)
-	s.Nil(encoding.TryParsingCustomError(err))
-	e := <-sink
-	s.NotNil(e)
-
-	difficulty, err := encoding.CalculatePacayaDifficulty(new(big.Int).SetUint64(e.Info.LastBlockId))
-	s.Nil(err)
-
-	parent, err := s.RPCClient.L2.HeaderByNumber(
-		context.Background(),
-		new(big.Int).Sub(new(big.Int).SetUint64(e.Info.LastBlockId), common.Big1),
-	)
-	s.Nil(err)
-
-	baseFee, err := s.RPCClient.CalculateBaseFee(
-		context.Background(), parent, &e.Info.BaseFeeConfig, e.Info.LastBlockTimestamp,
-	)
-	s.Nil(err)
-
-	anchorBlock, err := s.RPCClient.L1.HeaderByNumber(context.Background(), new(big.Int).SetUint64(e.Info.AnchorBlockId))
-	s.Nil(err)
-
-	anchorTxConstructor, err := anchortxconstructor.New(s.RPCClient)
-	s.Nil(err)
-	anchor, err := anchorTxConstructor.AssembleAnchorV3Tx(
-		context.Background(),
-		anchorBlock.Number,
-		anchorBlock.Root,
-		parent,
-		&e.Info.BaseFeeConfig,
-		[][32]byte{},
-		new(big.Int).SetUint64(e.Info.LastBlockId),
-		baseFee,
-	)
-	s.Nil(err)
-
-	txListBytes, err := rlp.EncodeToBytes(types.Transactions{anchor})
-	s.Nil(err)
-
-	s.forkTo(&engine.PayloadAttributes{
-		Timestamp:             e.Info.LastBlockTimestamp,
-		Random:                common.BytesToHash(difficulty),
-		SuggestedFeeRecipient: e.Info.Coinbase,
-		Withdrawals:           []*types.Withdrawal{},
-		BlockMetadata: &engine.BlockMetadata{
-			Beneficiary: e.Info.Coinbase,
-			GasLimit:    uint64(e.Info.GasLimit) + consensus.AnchorV3GasLimit,
-			Timestamp:   e.Info.LastBlockTimestamp,
-			TxList:      txListBytes,
-			MixHash:     common.Hash(difficulty),
-			ExtraData:   e.Info.ExtraData[:],
-		},
-		BaseFeePerGas: baseFee,
-		L1Origin: &rawdb.L1Origin{
-			BlockID:            new(big.Int).SetUint64(e.Info.LastBlockId),
-			L1BlockHeight:      new(big.Int).SetUint64(e.Raw.BlockNumber),
-			L2BlockHash:        common.Hash{},
-			L1BlockHash:        e.Raw.BlockHash,
-			BuildPayloadArgsID: [8]byte{},
-		},
-	}, parent.Hash())
-
-	head, err := s.RPCClient.L2.HeaderByNumber(context.Background(), nil)
-	s.Nil(err)
-	s.Equal(common.Big1, head.Number)
 }
